feat : 대기열 최신화 및 redis pub/sub을 사용한 메세징 처리
