+++ src/main/java/com/ureka/team3/utong_scheduler/config/SchedulingConfig.java
@@ -1,6 +1,7 @@
 package com.ureka.team3.utong_scheduler.config;
 
-import com.ureka.team3.utong_scheduler.contract.scheduler.ContractAggregationScheduler;
+import com.ureka.team3.utong_scheduler.trade.chart.scheduler.ContractAggregationScheduler;
+import com.ureka.team3.utong_scheduler.trade.chart.service.CurrentPriceService;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
@@ -12,9 +13,11 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 public class SchedulingConfig {
 
     private final ContractAggregationScheduler contractAggregationScheduler;
+    private final CurrentPriceService currentPriceService;
 
     @PostConstruct
     void init() {
         contractAggregationScheduler.insertInitialData();
+        currentPriceService.init();
     }
 }
