--- src/main/java/com/ureka/team3/utong_scheduler/contract/dto/AvgPerHour.java
@@ -1,31 +1,0 @@
-package com.ureka.team3.utong_scheduler.contract.dto;
-
-import com.ureka.team3.utong_scheduler.contract.entity.ContractHourlyAvgPrice;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Getter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class AvgPerHour {
-
-    private String dataCode;
-
-    private Long avgPrice;
-
-    private LocalDateTime aggregatedAt;
-
-    public static AvgPerHour of(ContractHourlyAvgPrice contractHourlyAvgPrice) {
-        AvgPerHour avgPerHour = new AvgPerHour();
-
-        avgPerHour.dataCode = contractHourlyAvgPrice.getDataCode();
-        avgPerHour.avgPrice = contractHourlyAvgPrice.getAvgPrice();
-        avgPerHour.aggregatedAt = contractHourlyAvgPrice.getAggregatedAt();
-
-        return avgPerHour;
-    }
-
-}
