--- src/main/java/com/ureka/team3/utong_scheduler/contract/service/CurrentPriceService.java
@@ -1,14 +1,0 @@
-package com.ureka.team3.utong_scheduler.contract.service;
-
-
-import com.ureka.team3.utong_scheduler.contract.dto.AvgPerHour;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface CurrentPriceService {
-
-    void updateRedisCache(String dataCode, LocalDateTime aggregatedAt);
-
-    List<AvgPerHour> getUpdatedData(String dataCode);
-}
