+++ src/main/java/com/ureka/team3/utong_scheduler/trade/queue/service/TradeQueueService.java
@@ -1,0 +1,16 @@
+package com.ureka.team3.utong_scheduler.trade.queue.service;
+
+
+import com.ureka.team3.utong_scheduler.trade.queue.dto.OrdersQueueDto;
+
+import java.util.Map;
+
+public interface TradeQueueService {
+
+    // <가격, 수량>
+    Map<Long, Long> getAllSellOrderNumbers(String dataCode);
+
+    Map<Long, Long> getAllBuyOrderNumbers(String dataCode);
+
+    void saveAllOrdersNumber(Map<String, OrdersQueueDto> dataMap);
+}
